use classes as suggested in PR review and move caches into its own file
